+++ src/post/service/rpost-like.service.ts
@@ -1,7 +1,7 @@
 import { Inject, Injectable, forwardRef } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { PostLike } from "../entity/post-like.entity";
-import { UserService } from "src/user/user.service";
+import { UserService } from "src/user/service/user.service";
 import { Repository } from "typeorm";
 import { RPostService } from "../service/rpost.service";
 import { ToggleLike } from "src/common/enum/toggle-like.enum";
